refactor(j2cl): split compose test doubles + rich-component tests out of monster class (#1270)
The ~5k-line J2clComposeSurfaceControllerTest is the monster class #1270
calls out. First cut, no behavior change:

- ComposeTestDoubles.java: the shared package-private doubles
  (FakeGateway, FakeView, FakeAttachmentTransport, CapturingDeltaFactory,
  FakeFactory) moved out verbatim so focused split classes can reuse them.
- J2clComposeRichComponentTest.java: the 6 rich-text component /
  annotation reply-submit tests (J-UI-5, #1083) moved into their own
  per-flow class.

Same 182 tests run and pass (6 in the new class + 176 in the trimmed
original); test sources compile clean.
